Adding a new HeaderCreator API, which allows custom headers and footers in CellTable. HeaderCreator has one method #buildHeader(Utility) which lets the user build the entire header (or footer) table.  The Helper class is used to add a row to the header table, render a Header instance into the header table, or enable column-specific event handling on an element.  Headers (which are basically a Cell with a static value) now use the same method as Cells to fire events.  We walk up the DOM until we find an attribute added by Utility.renderHeader(), then lookup the header based on the attribute.  You can add multiple headers per DOM table cell.
Helper#enableColumnHandlers(ElementBuilderBase, Column) associates the specified column with the Element created by the specified builder.  This allows you to specify which columns are sortable, and which Column to pass to the ColumnSortList when sorting occurs.  With this change, Columns are entirely superficial if you specify a CellTableBuilder and a HeaderCreator. The default implementation renders the header based on the Columns, but if you provide your own builders, you do not need to specify any Columns in CellTable.  This is very useful if you want to break the grid pattern in favor of something more complicated.

The DefaultHeaderCreator creates one header per Column, exactly as you would expect. It also supports the method #setSortIconStartOfLine(boolean), which allows you to specify whether the sort icon appears on the right or left side (in RTL safe verbiage). This was a widely requested feature, and something that is easy to do with a HeaderBuilder.

The example in Showcase shows a multi row header with a combination of sortable columns and non-sortable columns. It also takes advantage of the fact the columns are optional to show how users might maintain columns manually.
http://showcase3.jlabanca-testing.appspot.com/#!CwCustomDataGrid

Review at http://gwt-code-reviews.appspot.com/1499808


git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@10581 8db76d5a-ed1c-0410-87a9-c151d255dfc7