Integrating Peng's recent API changes to CellTableBuilder into HeaderCreator.  HeaderCreator now includes both the rendering code and the event handling logic, eliminating the HeaderCreator.Helper class completely.  This allows for a simpler implementation of HeaderCreator that only supports a single row, or an application specific implementation, either of which could be lighter weight than the default implementation.  The API is very similar to CellTableBuilder, but slightly simpler because there is no concept of row values in a header.  Note that there are no behavioral changes here, its just a rearrangement of API.
The one big question, which this change does not answer, is which class to rename: HeaderCreator or CellTableBuilder.  Personally, I think that both classes are much more builder-like than they were, so maybe we can just rename HeaderCreator to HeaderBuilder.  That would certainly be the easiest from an adoption standpoint, as CellTableBuilder already has wide use and would be harder to rename.

Also, we could drop the "Cell" from CellTableBuilder to make it more consistent with HeaderBuilder/Creator.  However, that would be a pain, so I'm happy just leaving CellTableBuilder as is.

Review at http://gwt-code-reviews.appspot.com/1533804


git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@10597 8db76d5a-ed1c-0410-87a9-c151d255dfc7