Update JDT from 3.4.2 to 3.4.2_r894.
This commit replaces the shadowed version of SourceTypeBinding.java with the original from 3.4.2_r894. We'll apply our modifications to this files in a subsequent commit (to make it easier to look back through code history and see what the changes that we made to the original are).

MethodBinding.java and BinaryTypeBinding.java have been removed because the changes that we made to the 3.4.2 originals have are no longer needed, or have been subsumed by API changes in 3.4.2_r894.

Review by: scottb@google.com

git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@10868 8db76d5a-ed1c-0410-87a9-c151d255dfc7