Remove dependency of Request Factory on ASM.
This is in preparation to the move to Maven, so that ASM as a whole is not
needed as a dependency only for a couple of method calls. This is why this
change doesn't affect RequestFactoryJarExtractor, which will go away with
the move to Maven.

Change-Id: I919e5d1a8f59d01d0de468815582e6cc0133cfbc
Review-Link: https://gwt-review.googlesource.com/#/c/1032/

Review by: skybrian@google.com

git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@11399 8db76d5a-ed1c-0410-87a9-c151d255dfc7