Use standards mode for xsiframe linker iframes when appropriate
When creating a child iframe, write out a standards mode doctype if
the host document is in standards mode.  In IE8, it's possible for the
host page to be in standards mode while the iframe child is in quirks
mode, which can cause unexpected failures like "$doc.querySelectorAll"
is undefined.

A simple GWTTestCase-compatible test for this is:

public void testQuirk() {
assertNotNull(getSelector());
}

private native JavaScriptObject getSelector() /*-{
return $doc.body.querySelectorAll;
}-*/;

This test currently fails in IE8 with xsiframe linker, but passes with
this patch when the host page is in standards mode.  (It still fails
in quirks mode, which I believe is to be expected.)

Change-Id: I1dfa37879412e335ed217307e5398dc4c897f785
Review-Link: https://gwt-review.googlesource.com/#/c/1310/


git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@11405 8db76d5a-ed1c-0410-87a9-c151d255dfc7