Window.Location.getParameter and getParameterMap now reflect changes to window history.
This is needed when using the HTML 5 window.history.pushState method, which modifies
the value of window.location without a browser reload.

This breaks the assumption in GWT's Window.Location class: window.location is not immutable anymore,
and therefore there needs to be a way to update the static variable Window.Location.listParamMap.

I looked at WindowTest but it seems it's not possible to change browser state in the test (in this case I would need to change the value of WindowImpl.getQueryString),
hence the lack of test in this patch.

Review at http://gwt-code-reviews.appspot.com/1859804
Fixes issue 7769

Review by: skybrian@google.com

git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@11419 8db76d5a-ed1c-0410-87a9-c151d255dfc7