Fixes GWT.runAsync that was broken in a recent patch.
This is a regression to my previous change to GWT.runAsync to make it always
run async.
In the previous patch, GWT.runAsync was calling overloaded GWT.runAsync that
is treated as a split point itself. Apparently, it was treated as a split 
point before the code was replaced for production mode.

This patch avoids the problem by delegating both runAsync calls to a helper
method.

Change-Id: If396a7f72ac533b304daf4d8afd8684d25e018ba