InternalCompilerError analyzing @SuppressWarnings in JsniChecker.
JsniChecker analyzes @SuppressWarnings annotations in order to allow
suppression of warnings from jsni code. The only requisite is that
it can only analyze constant string parameters as in this point of
compilation it can not chase variables.

With this fix, JsniChecker does not abort when unable to analyze the
annotation, just logs a warning.

Fixes issue 8093.

Change-Id: Ib42b75a363e3b293c90e4c6ae7c1bb0dd941bab5