Fixes potential infinite recursion in JavascriptException#getMessage.
This is a fix to regression introduced with unwrapping to JavaScriptExceptions (JSE).
The referred patch added a new field to wrapped JSO to point back to constructed JSE.
However, when not emulated stack traces are used, JSE prints all properties of the
wrapped JSO to the message of the exception. This causes following recursion:

  JSE#toString
    JSE#getMessage
      JSE#getExceptionProperies(JSO)
        JSO#__gwt$exception.toString
          JSE#toString (<same state as the begining of stack>)

This was not caught with test cases because eventually an exception was thrown and later
caught and ignored via JSE#getExceptionProperties.

With this patch, the recusion is avoided by adding __gwt$exception to special-cased
properties inside the StackTraceCreate.Collector.

Change-Id: I7ac8fead31208f49492f5bda1912051a41c289a3