Gets rid of long waits inside MediaSuite
MediaSuite test are slowest running test suite in GWT-SDK due to unnecessary
long waits inside the test cases. As it is slowest suie, it slows down whole
test execution when the tests are executed in parallel.

This patch changes the test pattern so that they are executed as fast as
possible via polling the state.

Suite now finishes ~1 min instead of ~3 min; i.e less wait for smoke test results!

Change-Id: I4a2a67884aea2d060343c4afaf3aff99ed48ab51