Fix CompoundAssignmentNormalizer to create locals with declarations.
A previous patch removed the creation of declaration statement nodes
in the AST during temp local creation. The current JavaScript
generation approach does not rely on the existence of these
declaration statement nodes since it uses the function scope locals
list, but the nodes might be needed in the future.

So this patch is restoring them.

Change-Id: I9a4cf9b4dda02f05ad705f94df8bded24fa8519d