Stopped routing localWorkers # through CompilerContext
When introducing CompilerContext an attempt was made to also remove
now-redundant option parameters, including "localWorkers". But
because of the swamp of option object hierarchy, that this one was
improper. Restoring this option parameter.

At some future date after an options cleanup it may be possible to
route this setting through CompilerContext.

Change-Id: I751a0f24774e8ba58ea92bf1b0b98383a8dd0bcb